panel-api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
google-cloud-storage==2.17.0
python-multipart==0.0.9
